ci(benchmark/react): enable list in benchmark (#1635)

+++ packages/lynx/benchx_cli/scripts/prepare.mjs
@@ -2,9 +2,16 @@
 // Licensed under the Apache License Version 2.0 that can be found in the
 // LICENSE file in the root directory of this source tree.
 
-import { copyFileSync, mkdirSync, writeFileSync } from 'node:fs';
+import { copyFileSync, existsSync, mkdirSync, writeFileSync } from 'node:fs';
 
 mkdirSync('./dist/bin', { recursive: true });
+
+if (existsSync('./dist/bin/benchx_cli')) {
+  // File already exists
+  // eslint-disable-next-line n/no-process-exit
+  process.exit(0);
+}
+
 if (process.platform === 'win32') {
   writeFileSync(
     './dist/bin/benchx_cli',
